#65 chore: LoginInfoArea(로그인 영역) 뷰 영역만 별도 컴포넌트로 분리
